pyproject.toml
[project]
name = "llama-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "llama-index>=0.14.18",
    "python-decouple>=3.8",
]
